Refactor generator to create server once
